% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise01-data.R
\docType{data}
\name{noise01}
\alias{noise01}
\title{Random 1 second of noise to preceed playback events}
\format{
A data.table of 200000 rows and 3 variables
}
\source{
2-3 seconds of "/home/jason/Documents/swarm_recordings/incubator_tests/20200110_G3_100_cage8/20200110_1630_eveningswarm_500Hzplayback_000.mat"
}
\usage{
data(noise01)
}
\description{
Random 1 second of noise to preceed playback events
}
\keyword{datasets}
